+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+## v1.0.2 (2023-08-21)
+
+- buat Models => php spark make:model ProductsModel
+- buat Controllers => php spark make:controller ProductsController --restful
+
 ## v1.0.1 (2023-08-21)
 
 - DATABASE Connection